ascii and word count check for shortest message
Check if string is not a special character (or an emoticon placeholder)
before confirming it as the shortest string.
Emoticon placeholders are processed as 4 different characters in some
cases, so it will be ignored if the message length <5 and the message is
non-ascii. It also checks if the message is more than 1 word. Feel free
to change in later commits.